add HostRestController(not test)
